refactor: refactor api for manage subject
